refactor(query): Extract tests to dedicated test files
Move test modules from parser.rs files to separate test files:
- sdl_parse/parser.rs tests → sdl_parse/parser_tests.rs (1,709 lines)
- query_parse/parser.rs tests → mutation_tests.rs, variable_tests.rs, subscription_tests.rs

Reduces sdl_parse/parser.rs from 1,989 → 280 lines.
Reduces query_parse/parser.rs from 2,240 → 1,166 lines.
