+++ crates/query/src/query_parse/parser.rs
@@ -1154,1087 +1154,13 @@ fn resolve_bool_value(
 }
 
 #[cfg(test)]
-mod mutation_tests {
-    use super::*;
-    use crate::mapper::MutationType;
-
-    #[test]
-    fn test_parse_create_mutation() {
-        let query = r#"
-            mutation {
-                create_Users(input: [{name: "Alice", age: 30}]) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        assert_eq!(mutations.len(), 1);
-
-        let m = &mutations[0];
-        assert_eq!(m.mutation_type, MutationType::Create);
-        assert_eq!(m.collection_name, "Users");
-        assert_eq!(m.create_input.len(), 1);
-        assert_eq!(
-            m.create_input[0].get("name"),
-            Some(&JsonValue::String("Alice".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_parse_create_multiple_documents() {
-        let query = r#"
-            mutation {
-                create_Users(input: [
-                    {name: "Alice", age: 30},
-                    {name: "Bob", age: 25}
-                ]) {
-                    _docID
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        assert_eq!(mutations[0].create_input.len(), 2);
-    }
-
-    #[test]
-    fn test_parse_update_mutation() {
-        let query = r#"
-            mutation {
-                update_Users(docIDs: ["bae-123"], input: {email: "new@example.com"}) {
-                    _docID
-                    email
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        assert_eq!(mutations.len(), 1);
-
-        let m = &mutations[0];
-        assert_eq!(m.mutation_type, MutationType::Update);
-        assert_eq!(m.collection_name, "Users");
-        assert_eq!(m.doc_ids, Some(vec!["bae-123".to_string()]));
-        assert_eq!(
-            m.update_input.get("email"),
-            Some(&JsonValue::String("new@example.com".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_parse_update_with_filter() {
-        let query = r#"
-            mutation {
-                update_Users(filter: {name: {_eq: "Alice"}}, input: {active: false}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        let m = &mutations[0];
-        assert!(m.filter.is_some());
-        assert!(m.doc_ids.is_none());
-    }
-
-    #[test]
-    fn test_parse_delete_mutation() {
-        let query = r#"
-            mutation {
-                delete_Users(docIDs: ["bae-123", "bae-456"]) {
-                    _docID
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        assert_eq!(mutations.len(), 1);
-
-        let m = &mutations[0];
-        assert_eq!(m.mutation_type, MutationType::Delete);
-        assert_eq!(m.collection_name, "Users");
-        assert_eq!(
-            m.doc_ids,
-            Some(vec!["bae-123".to_string(), "bae-456".to_string()])
-        );
-    }
-
-    #[test]
-    fn test_parse_delete_with_filter() {
-        let query = r#"
-            mutation {
-                delete_Users(filter: {active: {_eq: false}}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        let m = &mutations[0];
-        assert!(m.filter.is_some());
-    }
-
-    #[test]
-    fn test_parse_multiple_mutations() {
-        let query = r#"
-            mutation {
-                create_Users(input: [{name: "Alice"}]) {
-                    _docID
-                }
-                delete_Posts(docIDs: ["bae-999"]) {
-                    _docID
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        assert_eq!(mutations.len(), 2);
-        assert_eq!(mutations[0].mutation_type, MutationType::Create);
-        assert_eq!(mutations[1].mutation_type, MutationType::Delete);
-    }
-
-    #[test]
-    fn test_create_missing_input_error() {
-        let query = r#"
-            mutation {
-                create_Users {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("requires 'input'"));
-    }
-
-    #[test]
-    fn test_update_without_target_succeeds() {
-        // Go DefraDB allows update without filter or docIDs (meaning update all)
-        let query = r#"
-            mutation {
-                update_Users(input: {name: "Bob"}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(
-            result.is_ok(),
-            "update without target should succeed: {:?}",
-            result
-        );
-        let mutations = result.unwrap();
-        assert_eq!(mutations.len(), 1);
-        assert!(mutations[0].doc_ids.is_none());
-        assert!(mutations[0].filter.is_none());
-    }
-
-    #[test]
-    fn test_delete_without_target_succeeds() {
-        // Go DefraDB allows delete without filter or docIDs (meaning delete all)
-        let query = r#"
-            mutation {
-                delete_Users {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(
-            result.is_ok(),
-            "delete without target should succeed: {:?}",
-            result
-        );
-        let mutations = result.unwrap();
-        assert_eq!(mutations.len(), 1);
-        assert!(mutations[0].doc_ids.is_none());
-        assert!(mutations[0].filter.is_none());
-    }
-
-    #[test]
-    fn test_invalid_mutation_name_error() {
-        let query = r#"
-            mutation {
-                Users(input: [{name: "Alice"}]) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("Invalid mutation name"));
-    }
-
-    #[test]
-    fn test_query_still_works() {
-        let query = r#"
-            {
-                Users {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let selects = parse_query(query).unwrap();
-        assert_eq!(selects.len(), 1);
-        assert_eq!(selects[0].collection_name, "Users");
-    }
-
-    #[test]
-    fn test_cannot_mix_query_and_mutation() {
-        // Note: GraphQL parser won't actually allow this syntax,
-        // but we handle it anyway
-        let query = r#"
-            mutation {
-                create_Users(input: [{name: "Alice"}]) { _docID }
-            }
-        "#;
-
-        // This should work as pure mutation
-        let result = parse_mutations(query);
-        assert!(result.is_ok());
-
-        // parse_query should fail on mutation
-        let result = parse_query(query);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_upsert_mutation_go_style() {
-        // Go DefraDB upsert syntax: filter, create, update (all required)
-        let query = r#"
-            mutation {
-                upsert_Users(
-                    filter: {name: {_eq: "Bob"}},
-                    create: {name: "Bob", age: 40},
-                    update: {age: 40}
-                ) {
-                    _docID
-                    name
-                    age
-                }
-            }
-        "#;
-
-        let mutations = parse_mutations(query).unwrap();
-        assert_eq!(mutations.len(), 1);
-
-        let m = &mutations[0];
-        assert_eq!(m.mutation_type, MutationType::Upsert);
-        assert_eq!(m.collection_name, "Users");
-        assert!(m.filter.is_some());
-        // create_input is stored as a single-element vec
-        assert_eq!(m.create_input.len(), 1);
-        assert_eq!(
-            m.create_input[0].get("name"),
-            Some(&JsonValue::String("Bob".to_string()))
-        );
-        // update_input is the fields to update
-        assert_eq!(
-            m.update_input.get("age"),
-            Some(&JsonValue::Number(40.into()))
-        );
-    }
-
-    #[test]
-    fn test_upsert_missing_filter_error() {
-        // Go style requires filter
-        let query = r#"
-            mutation {
-                upsert_Users(
-                    create: {name: "Bob", age: 40},
-                    update: {age: 40}
-                ) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("filter"));
-    }
-
-    #[test]
-    fn test_upsert_missing_create_error() {
-        // Go style requires create
-        let query = r#"
-            mutation {
-                upsert_Users(
-                    filter: {name: {_eq: "Bob"}},
-                    update: {age: 40}
-                ) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("create"));
-    }
-
-    #[test]
-    fn test_upsert_missing_update_error() {
-        // Go style requires update
-        let query = r#"
-            mutation {
-                upsert_Users(
-                    filter: {name: {_eq: "Bob"}},
-                    create: {name: "Bob", age: 40}
-                ) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_mutations(query);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("update"));
-    }
-}
+#[path = "mutation_tests.rs"]
+mod mutation_tests;
 
 #[cfg(test)]
-mod variable_tests {
-    use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn test_variable_in_filter() {
-        let query = r#"
-            query($name: String!) {
-                Users(filter: {name: {_eq: $name}}) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("name".to_string(), json!("Alice"))]);
-
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(selects.len(), 1);
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                let name_cond = conditions.get("name").unwrap();
-                assert_eq!(name_cond.get("_eq"), Some(&json!("Alice")));
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_in_limit() {
-        let query = r#"
-            query($lim: Int!) {
-                Users(limit: $lim) {
-                    _docID
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("lim".to_string(), json!(10))]);
-
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(selects[0].limit.as_ref().unwrap().limit, Some(10));
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_in_doc_ids() {
-        let query = r#"
-            query($ids: [String!]!) {
-                Users(docIDs: $ids) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("ids".to_string(), json!(["bae-123", "bae-456"]))]);
-
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(
-                    selects[0].doc_ids,
-                    Some(vec!["bae-123".to_string(), "bae-456".to_string()])
-                );
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_in_mutation_input() {
-        let query = r#"
-            mutation($userName: String!, $userAge: Int!) {
-                create_Users(input: [{name: $userName, age: $userAge}]) {
-                    _docID
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([
-            ("userName".to_string(), json!("Bob")),
-            ("userAge".to_string(), json!(25)),
-        ]);
-
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Mutation { mutations, .. } => {
-                assert_eq!(mutations.len(), 1);
-                let input = &mutations[0].create_input[0];
-                assert_eq!(input.get("name"), Some(&json!("Bob")));
-                assert_eq!(input.get("age"), Some(&json!(25)));
-            }
-            _ => panic!("Expected mutation"),
-        }
-    }
-
-    #[test]
-    fn test_variable_in_mutation_doc_ids() {
-        let query = r#"
-            mutation($id: String!) {
-                delete_Users(docIDs: [$id]) {
-                    _docID
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("id".to_string(), json!("bae-999"))]);
-
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Mutation { mutations, .. } => {
-                assert_eq!(mutations[0].doc_ids, Some(vec!["bae-999".to_string()]));
-            }
-            _ => panic!("Expected mutation"),
-        }
-    }
-
-    #[test]
-    fn test_undefined_variable_error() {
-        let query = r#"
-            query {
-                Users(filter: {name: {_eq: $undefined}}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let variables = HashMap::new();
-        let result = parse_request_with_variables(query, Some(&variables), None);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("was not provided"));
-    }
-
-    #[test]
-    fn test_no_variables_provided_error() {
-        let query = r#"
-            query {
-                Users(filter: {name: {_eq: $name}}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_request_with_variables(query, None, None);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("no variables provided"));
-    }
-
-    #[test]
-    fn test_query_without_variables_still_works() {
-        let query = r#"
-            {
-                Users(filter: {name: {_eq: "Alice"}}) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        // No variables provided
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(selects.len(), 1);
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                let name_cond = conditions.get("name").unwrap();
-                assert_eq!(name_cond.get("_eq"), Some(&json!("Alice")));
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_type_mismatch_int() {
-        let query = r#"
-            query($lim: Int!) {
-                Users(limit: $lim) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Provide string instead of int
-        let variables = HashMap::from([("lim".to_string(), json!("not an int"))]);
-        let result = parse_request_with_variables(query, Some(&variables), None);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("must be of type Int"));
-    }
-
-    #[test]
-    fn test_multiple_variables() {
-        let query = r#"
-            query($name: String!, $minAge: Int!, $lim: Int!) {
-                Users(filter: {name: {_eq: $name}, age: {_gte: $minAge}}, limit: $lim) {
-                    _docID
-                    name
-                    age
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([
-            ("name".to_string(), json!("Alice")),
-            ("minAge".to_string(), json!(18)),
-            ("lim".to_string(), json!(5)),
-        ]);
-
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(selects[0].limit.as_ref().unwrap().limit, Some(5));
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                assert_eq!(
-                    conditions.get("name").unwrap().get("_eq"),
-                    Some(&json!("Alice"))
-                );
-                assert_eq!(conditions.get("age").unwrap().get("_gte"), Some(&json!(18)));
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    // =========================================================================
-    // Variable type mismatch tests
-    // =========================================================================
-
-    #[test]
-    fn test_variable_type_mismatch_bool() {
-        let query = r#"
-            query($deleted: Boolean!) {
-                Users(showDeleted: $deleted) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Provide string instead of bool
-        let variables = HashMap::from([("deleted".to_string(), json!("true"))]);
-        let result = parse_request_with_variables(query, Some(&variables), None);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("must be of type Boolean"));
-    }
-
-    #[test]
-    fn test_variable_type_mismatch_string() {
-        let query = r#"
-            query($c: String!) {
-                Users(cid: $c) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Provide integer instead of string
-        let variables = HashMap::from([("c".to_string(), json!(12345))]);
-        let result = parse_request_with_variables(query, Some(&variables), None);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("must be of type String"));
-    }
-
-    #[test]
-    fn test_variable_in_order_direction() {
-        let query = r#"
-            query($dir: String!) {
-                Users(order: {name: $dir}) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("dir".to_string(), json!("DESC"))]);
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                let order = selects[0].order_by.as_ref().unwrap();
-                assert_eq!(
-                    order.conditions[0].direction,
-                    crate::mapper::OrderDirection::Desc
-                );
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_invalid_order_direction() {
-        let query = r#"
-            query($dir: String!) {
-                Users(order: {name: $dir}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("dir".to_string(), json!("INVALID"))]);
-        let result = parse_request_with_variables(query, Some(&variables), None);
-        assert!(result.is_err());
-        // Error format matches Go DefraDB
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("Argument \"order\" has invalid value"));
-    }
-
-    // =========================================================================
-    // Variable default value tests
-    // =========================================================================
-
-    #[test]
-    fn test_variable_default_value_used_when_not_provided() {
-        let query = r#"
-            query($name: String = "DefaultName") {
-                Users(filter: {name: {_eq: $name}}) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                assert_eq!(
-                    conditions.get("name").unwrap().get("_eq"),
-                    Some(&json!("DefaultName"))
-                );
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_provided_value_overrides_default() {
-        let query = r#"
-            query($name: String = "DefaultName") {
-                Users(filter: {name: {_eq: $name}}) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        // Provide a value - should override default
-        let variables = HashMap::from([("name".to_string(), json!("ProvidedName"))]);
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                assert_eq!(
-                    conditions.get("name").unwrap().get("_eq"),
-                    Some(&json!("ProvidedName"))
-                );
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_default_int_value() {
-        let query = r#"
-            query($lim: Int = 50) {
-                Users(limit: $lim) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default 50
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(selects[0].limit.as_ref().unwrap().limit, Some(50));
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_default_boolean_value() {
-        let query = r#"
-            query($deleted: Boolean = true) {
-                Users(showDeleted: $deleted) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default true
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert!(selects[0].show_deleted);
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_multiple_variables_with_some_defaults() {
-        let query = r#"
-            query($name: String!, $minAge: Int = 18, $lim: Int = 10) {
-                Users(filter: {name: {_eq: $name}, age: {_gte: $minAge}}, limit: $lim) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        // Only provide $name, use defaults for $minAge and $lim
-        let variables = HashMap::from([("name".to_string(), json!("Alice"))]);
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                assert_eq!(selects[0].limit.as_ref().unwrap().limit, Some(10));
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                assert_eq!(
-                    conditions.get("name").unwrap().get("_eq"),
-                    Some(&json!("Alice"))
-                );
-                assert_eq!(conditions.get("age").unwrap().get("_gte"), Some(&json!(18)));
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_mutation_variable_default_value() {
-        let query = r#"
-            mutation($name: String = "DefaultUser") {
-                create_Users(input: [{name: $name}]) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Mutation { mutations, .. } => {
-                let input = &mutations[0].create_input;
-                assert_eq!(input[0].get("name"), Some(&json!("DefaultUser")));
-            }
-            _ => panic!("Expected mutation"),
-        }
-    }
-
-    #[test]
-    fn test_variable_default_array_value() {
-        let query = r#"
-            query($ids: [String!] = ["id1", "id2"]) {
-                Users(docIDs: $ids) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default array
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                let doc_ids = selects[0].doc_ids.as_ref().unwrap();
-                assert_eq!(doc_ids.len(), 2);
-                assert_eq!(doc_ids[0], "id1");
-                assert_eq!(doc_ids[1], "id2");
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_default_null_value() {
-        let query = r#"
-            query($name: String = null) {
-                Users(filter: {name: {_eq: $name}}) {
-                    _docID
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default null
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Query { selects, .. } => {
-                let filter = selects[0].filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                assert_eq!(
-                    conditions.get("name").unwrap().get("_eq"),
-                    Some(&JsonValue::Null)
-                );
-            }
-            _ => panic!("Expected query"),
-        }
-    }
-
-    #[test]
-    fn test_variable_default_cannot_reference_other_variable() {
-        // Note: GraphQL spec doesn't allow variable references in default values
-        // graphql-parser rejects this at parse time with a parse error
-        let query = r#"
-            query($a: String = $b, $b: String = "test") {
-                Users(filter: {name: {_eq: $a}}) {
-                    _docID
-                }
-            }
-        "#;
-
-        let result = parse_request_with_variables(query, None, None);
-        // graphql-parser rejects this at the parse level since variable references
-        // aren't allowed in default value position per the GraphQL spec
-        assert!(
-            result.is_err(),
-            "Expected error for variable reference in default value, but got: {:?}",
-            result
-        );
-    }
-}
+#[path = "variable_tests.rs"]
+mod variable_tests;
 
 #[cfg(test)]
-mod subscription_tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_subscription_basic() {
-        let query = r#"
-            subscription {
-                User {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let result = parse_request(query).unwrap();
-        match result {
-            ParsedOperation::Subscription { select } => {
-                assert_eq!(select.collection_name, "User");
-                assert_eq!(select.fields.len(), 2);
-            }
-            _ => panic!("Expected subscription"),
-        }
-    }
-
-    #[test]
-    fn test_parse_subscription_with_filter() {
-        let query = r#"
-            subscription {
-                User(filter: {active: {_eq: true}}) {
-                    _docID
-                    name
-                    email
-                }
-            }
-        "#;
-
-        let result = parse_request(query).unwrap();
-        match result {
-            ParsedOperation::Subscription { select } => {
-                assert_eq!(select.collection_name, "User");
-                assert!(select.filter.is_some());
-            }
-            _ => panic!("Expected subscription"),
-        }
-    }
-
-    #[test]
-    fn test_parse_subscription_with_variables() {
-        let query = r#"
-            subscription($active: Boolean!) {
-                User(filter: {active: {_eq: $active}}) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let variables = HashMap::from([("active".to_string(), serde_json::json!(true))]);
-        let result = parse_request_with_variables(query, Some(&variables), None).unwrap();
-        match result {
-            ParsedOperation::Subscription { select } => {
-                assert_eq!(select.collection_name, "User");
-                let filter = select.filter.as_ref().unwrap();
-                let conditions = filter.conditions();
-                assert_eq!(
-                    conditions.get("active").unwrap().get("_eq"),
-                    Some(&serde_json::json!(true))
-                );
-            }
-            _ => panic!("Expected subscription"),
-        }
-    }
-
-    #[test]
-    fn test_parse_subscription_multiple_root_fields_error() {
-        let query = r#"
-            subscription {
-                User {
-                    name
-                }
-                Post {
-                    title
-                }
-            }
-        "#;
-
-        let result = parse_request(query);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("exactly one root field"));
-    }
-
-    #[test]
-    fn test_parse_subscription_with_nested_fields() {
-        let query = r#"
-            subscription {
-                User {
-                    _docID
-                    name
-                    posts {
-                        _docID
-                        title
-                    }
-                }
-            }
-        "#;
-
-        let result = parse_request(query).unwrap();
-        match result {
-            ParsedOperation::Subscription { select } => {
-                assert_eq!(select.collection_name, "User");
-                // Should have 3 fields: _docID, name, and posts (nested)
-                assert_eq!(select.fields.len(), 3);
-            }
-            _ => panic!("Expected subscription"),
-        }
-    }
-
-    #[test]
-    fn test_cannot_mix_subscription_and_query() {
-        // GraphQL doesn't allow mixing operation types in a single document
-        // But we can test that our parser handles it correctly
-        let query = r#"
-            subscription {
-                User { name }
-            }
-        "#;
-
-        let query_op = r#"
-            query {
-                User { name }
-            }
-        "#;
-
-        // Each should parse independently
-        assert!(matches!(
-            parse_request(query).unwrap(),
-            ParsedOperation::Subscription { .. }
-        ));
-        assert!(matches!(
-            parse_request(query_op).unwrap(),
-            ParsedOperation::Query { .. }
-        ));
-    }
-
-    #[test]
-    fn test_subscription_with_doc_id() {
-        let query = r#"
-            subscription {
-                User(docID: "bae-123") {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        let result = parse_request(query).unwrap();
-        match result {
-            ParsedOperation::Subscription { select } => {
-                assert_eq!(select.doc_ids, Some(vec!["bae-123".to_string()]));
-            }
-            _ => panic!("Expected subscription"),
-        }
-    }
-
-    #[test]
-    fn test_subscription_with_default_variable() {
-        let query = r#"
-            subscription($limit: Int = 10) {
-                User(limit: $limit) {
-                    _docID
-                    name
-                }
-            }
-        "#;
-
-        // Don't provide the variable - should use default
-        let result = parse_request_with_variables(query, None, None).unwrap();
-        match result {
-            ParsedOperation::Subscription { select } => {
-                assert_eq!(select.limit.as_ref().unwrap().limit, Some(10));
-            }
-            _ => panic!("Expected subscription"),
-        }
-    }
-}
+#[path = "subscription_tests.rs"]
+mod subscription_tests;
